Splitted MainHeader into DrawerTemplate and Header
